requirements.txt
aiogram
aiohttp
python-dotenv
qrcode[pil]
Pillow
